pdf-saas/json-to-pdf-convert-microservice/requirements.txt
fastapi
uvicorn
python-docx
python-multipart
python-dotenv

